gradio_deploy/requirements.txt
gradio
torch
torchvision
Pillow
numpy==1.26.0
matplotlib
boto3